+++ demo/js/directives.js
@@ -174,6 +174,9 @@
 			$http.get("http://www.imdb.com/xml/find?json=1&nr=1&tt=on&q=l");
 		};
 	}]);
-
+	angular.module('demoApp')
+	.controller('ibanCtrl', ['$scope', function($scope){
+		$scope.iban = "FR7630004000031234567890143";
+	}]);
 
 })();